\name{cdfenv}
\Rdversion{0.99.0}
\alias{cdfenv}

\title{
R environement for GEO series GSE19183
}
\description{
The \code{cdfenv} environment is a hash table for the annotation of the Affymetrix Genechip miRNA-1.0 used in the GEO series GSE19183.
}

\details{
This \code{cdfenv} environment is based on a CDF annotation file provided directly by Affymetrix, as explained
in the vignette of the \code{ExiMiR} package. It has been generated by the \code{make.cdf.env} function from the package \code{makecdfenv}.
}