\name{ExiMiR-package}
\alias{ExiMiR}
\docType{package}

\title{
R functions for the normalization of Exiqon miRNA array data
}

\description{
This package contains functions for reading raw data in ImaGene TXT format obtained from Exiqon miRCURY LNA arrays, annotating them with appropriate GAL files, and normalizing them using a spike-in probe-based method. Other platforms and data formats are also supported by ExiMiR.
}
\details{
\tabular{ll}{
Package: \tab ExiMiR\cr
Type: \tab Package\cr
Version: \tab 1.99.0\cr
Date: \tab 2012-04-24\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sylvain Gubian, Alain Sewer, PMP SA\cr
Maintainer: DL.RSupport@pmi.com
}
