\name{galenv}
\Rdversion{0.99.0}
\alias{galenv}

\title{
R environement for GEO series GSE20122
}
\description{
The \code{galenv} environment is a hash table for the annotation of the Exiqon miRCURY LNA arrays used in the GEO series GSE20122.
}

\details{
This \code{galenv} environment is based on a GAL annotation file provided directly by Exiqon, as explained
in the vignette of the \code{ExiMiR} package. It has been generated by the \code{make.gal.env} function from \code{ExiMiR}.
}