\name{GSE19183}
\Rdversion{0.99.0}
\alias{GSE19183}

\title{
Affybatch object for raw data from GEO series GSE19183
}
\description{
The Affybatch object \code{GSE19183} contains the raw expression data obtained from the CEL files of the GEO series GSE19183.
}
\details{
The Affybatch object \code{GSE19183} has been generated using the \code{ReadAffy} function from the package \code{affy} and its annotation is provided
by the \code{cdfenv} environment contained in \code{ExiMiR}.
}