\name{GSE20122}
\Rdversion{0.99.0}
\alias{GSE20122}

\title{
Affybatch object for raw data from GEO series GSE20122
}
\description{
The Affybatch object \code{GSE20122} contains the raw expression data obtained from the TXT files of the GEO series GSE20122 performed on
the Exiqon miRCURY platform.
}
\details{
The Affybatch object \code{GSE20122} has been generated using the \code{ReadExi} function from the package \code{ExiMiR} and its annotation is provided
by the \code{galenv} environment contained in \code{ExiMiR} as well.
}