\name{ReadExi}
\Rdversion{0.99.0}
\alias{ReadExi}
\title{
Exiqon 'txt' files reader
}
\description{
This function reads Exiqon 'txt' files and create an \code{AffyBatch} object.
}
\usage{
ReadExi(	txtfile.path= getwd(),
			galname= NULL,
			description = NULL,
			notes = '',
			rm.background = FALSE,
			verbose=TRUE)
}

\arguments{
  \item{txtfile.path}{
  Character vector. Path to the folder which contains samplesinfo.txt and Exiqon 'txt' files
}
  \item{galname}{
  Character vector. Name of a GAL environment generated by the ExiMiR \code{make.gal.env} function.
}
  \item{description}{
  a \code{MIAME} object.
}
  \item{notes}{
  notes.
}
  \item{rm.background}{
  Logical. Default value is \code{FALSE}. If \code{TRUE}, the background median is subtracted from the signal median.
}
  \item{verbose}{
  Logical. Default value is \code{TRUE}; some details are provided on the console
}
}
\details{
Exiqon 'txt' files are supplied with a samplesinfo.txt description file which lists the names of the samples files per channel.
The \code{txtfile.path} argument should be a folder that contains 'txt' files and a samplesinfo.txt file. If not, the \code{ReadExi} function stops.
The \code{galname} argument should be the name of the GAL environment created with \code{make.gal.env} function. If \code{galname} is not provided, an hashed environment is 
created based on the annotation that the 'txt' file contains.
}

\value{
An \code{AffyBatch} object.
}

\examples{
	# The folder 'Exiqon' contains the file 'samplesinfo.txt' and 
	# the corresponding raw data files in TXT format
	\dontrun{ebatch <- ReadExi(txtfile.path='Exiqon')}
	# If the GAL environment has already created by the function make.gal.env
	\dontrun{ebatch <- ReadExi(galenv='galenv, 'txtfile.path='Exiqon')}
}


\author{
Sylvain Gubian, Alain Sewer, PMP SA
}

\section{Warning}{
The \code{image} method from the \code{AffyBatch} object might not work properly when the \code{galname} argument is not assigned.
}

\seealso{
\code{AffyBatch}, \code{make.gal.env}
}
