\name{ExiMiR-package}
\Rdversion{0.99.0}
\alias{ExiMiR}
\docType{package}
\title{
R functions for the normalization of Exiqon miRNA array data
}
\description{
This package contains functions for reading raw data in ImaGene TXT format
obtained from Exiqon miRCURY LNA arrays, annotating them with appropriate GAL files,
and normalizing them using a spike-in probe-based method.
}
\details{
\tabular{ll}{
Package: \tab ExiMiR\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2010-10-29\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sylvain Gubian, Alain Sewer, PMP SA\cr
Maintainer: DL.RSupport@pmi.com
}
