\name{galenv}
\alias{galenv}

\title{
R annotation environment for GEO series GSE20122
}
\description{
The \code{galenv} environment is a hash table for the annotation of the Exiqon miRCURY
LNA arrays used in the GEO series GSE20122 (Exiqon miRCURY LNA array v.11).
}

\details{
See accompanying vignette.  
}

\seealso{
\code{make.gal.env}.
}
