\name{GSE19183}
\alias{GSE19183}

\title{
Affybatch object for the raw data from GEO series GSE19183
}
\description{
The Affybatch object \code{GSE19183} contains the raw expression data contained in the CEL files
of the GEO series GSE19183, obtained from the Affymetrix miRNA-1_0 platform. The annotation is
included in the Affybatch object. 
}
\details{
See accompanying vignette.  
}