\name{GSE20122}
\alias{GSE20122}

\title{
Affybatch object for the raw data from GEO series GSE20122
}
\description{
The Affybatch object \code{GSE20122} contains the raw expression data contained in the ImaGene TXT
files of the GEO series GSE20122, obtained from the Exiqon miRCURY LNA platform v.11.
}
\details{
See accompanying vignette.  
}